% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddQ.R
\name{AddQ}
\alias{AddQ}
\title{AddQ}
\usage{
AddQ(x, quantiles)
}
\arguments{
\item{x, }{a numeric vector}
}
\value{
The function return a vector oif labels
}
\description{
This function is to label the numberic vector by quantile
}
