% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbleplot.R
\name{bubblePrep}
\alias{bubblePrep}
\title{bubblePrep}
\usage{
bubblePrep(datapair, cpcol)
}
\arguments{
\item{cpcol}{The column name for comparison, the default is name}

\item{pair.data}{the data pair that was prepred by datapair.mk}
}
\value{
return bubble plot for plotting
}
\description{
This function is the internal function to generate bubble data for plotting
}
