% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_RePACT.r
\name{CallT2Dpeak_pvalueOneTail}
\alias{CallT2Dpeak_pvalueOneTail}
\title{CallT2Dpeak_pvalueOneTail}
\usage{
CallT2Dpeak_pvalueOneTail(
  cellvsPeak.m.aggr,
  depths,
  index,
  doscale = T,
  GlobalSlopes
)
}
\description{
CallT2Dpeak_pvalueOneTail
}
\examples{
PCAInfo.20bin.donor.PCA <- CallT2Dpeak_pvalueOneTail(PCAInfo.20bin.donor.ob$cellvsPeak.m.aggr, PCAInfo.20bin.donor.ob$depths, PCAInfo.20bin.donor.ob$index,doscale=T, GlobalSlopes=PCAInfo.20bin.ob.PCA$pseudoregress.all[,1,drop=F])
}
