% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_RePACT.r, R/snATAC.RePACT.r
\name{CallT2Dpeak_qvalue_SpeedUP}
\alias{CallT2Dpeak_qvalue_SpeedUP}
\title{CallT2Dpeak_qvalue_SpeedUP
This function is to run regression (linear/logistic) based on the number of PCs, and characteristics of samples.}
\usage{
CallT2Dpeak_qvalue_SpeedUP(
  cellvsPeak.m.aggr = ATAConCCA.betaT2D.tjct.3nd.10bin.ob$cellvsPeak.m.aggr,
  depths = ATAConCCA.betaT2D.tjct.3nd.10bin.ob$depths,
  index = ATAConCCA.betaT2D.tjct.3nd.10bin.ob$index,
  qcut = 0.1,
  slopecut1 = 0.5,
  slopecut2 = -0.5,
  doscale = T
)

CallT2Dpeak_qvalue_SpeedUP(
  cellvsPeak.m.aggr = ATAConCCA.betaT2D.tjct.3nd.10bin.ob$cellvsPeak.m.aggr,
  depths = ATAConCCA.betaT2D.tjct.3nd.10bin.ob$depths,
  index = ATAConCCA.betaT2D.tjct.3nd.10bin.ob$index,
  qcut = 0.1,
  slopecut1 = 0.5,
  slopecut2 = -0.5,
  doscale = T
)
}
\arguments{
\item{cellvsPeak.m.aggr, }{a vector of peaks (":" "-")}

\item{depths}{total fragments of cells within each traj_bin}

\item{index}{avg pseudo-index of each bin}

\item{qcut}{qvalue cutoff to measure the significance of glm(expr~index)}

\item{slopecut1}{UP slope cutoff, slope from glm(expr~index), UP genes}

\item{slopecut2}{DN slope cutoff, slope from glm(expr~index), DN genes}

\item{doscale}{if build glm model based on scaled_expression and scaled pseudo-inddex}

\item{depths, }{RePACT reesult OBJ}

\item{index, }{RePACT reesult OBJ}

\item{qcut, }{qvalue cutoff, default qcut=0.1}

\item{slopecut1, }{slope cutoff for UP-regulated genes/peaks, default slopecut1=0.5}

\item{slopecut2, }{slope cutoff for DN-regulated genes/peaks,default slopecut2=-0.5}

\item{doscale, }{if the regression is performed on scaled index and expr data, default doscale=T}
}
\value{
The function return a list: "PCvariance", "PCanfpheno", "object.raw.withinfo", "model", "reg.plot.2d", "model.para"

The function return a list of plots, pseudoregress.all contains summary statistics for all genes/peaks,
}
\description{
This function is to call dynamic peaks along the phenotype trajectory
}
\examples{
betaT2D.diffGene.20bin.PCA <- CallT2Dpeak_qvalue_SpeedUP(beta.RNA.PCA.20bin.ob$cellvsPeak.m.aggr, depths=beta.RNA.PCA.20bin.ob$depths, index=beta.RNA.PCA.20bin.ob$index, qcut=0.2,slopecut1=0.3,slopecut2=-0.3,doscale=T)
}
