% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{datapair.mk}
\alias{datapair.mk}
\title{datapair.mk}
\usage{
datapair.mk(Seurat.ob.list, cols, pick.list, normalizecellsize = F,
  randomizecelloirder = T)
}
\arguments{
\item{Seurat.ob.list}{a list of seurat object that I will extract data from}

\item{cols}{a vector of columns that will be considered for each object in order}

\item{pick.list}{a list of vectors that set the rules to take data}

\item{normalizecellsize, }{if true, then normalize the matrix by total cell umi. Default is F
'@param randomizecelloirder If T, randomize the cell order in matrix, default is T}
}
\value{
return data pair that can be used for DE, bubble plot and others.
}
\description{
This function is to make a datapair that includes UMI/normalized UMI ,atrix as well as information table, out of one or multiple seurat objects
}
\examples{
ROCKvsnorock.endo.paired<-datapair.mk(list(S7rock=S7rock_1.ob,S7=S7.ob),cols=c("Sample","Sample.2nd"),pick.list=list(c("s7.RockII_1"),c("s7.B_1")),normalizecellsize=F,randomizecelloirder=T)
}
