% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{dgepreprocess}
\alias{dgepreprocess}
\title{dgepreprocess}
\usage{
dgepreprocess(isletdge, Txcutoff = 1000, norowname = T)
}
\arguments{
\item{isletdge}{The raw dge data}

\item{Txcutoff}{The UMI cutoff as for defined "cell"}

\item{norowname}{The input dge file with or without row names? if true, input already has rowname(usually when data taken from other analyzed Seurat object), if FALSE, the input without rowname and ususally is the very raw data}
}
\value{
this will return the formated dge for clustering
}
\description{
This function is to preprocess raw dge into a format for clustering analysis, including add rownames and cutoff low expressing cells
}
\examples{
dgepreprocess(s7.RockII_1.dge,500,norowname=T)
}
