% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreCCAtree.R
\name{dist.matrix.prep.v3}
\alias{dist.matrix.prep.v3}
\title{dist.matrix.prep.v3}
\usage{
dist.matrix.prep.v3(binary.primary, datainfo.col1 = c("res.0.6", "nUMI",
  "Sample"), datainfo.col2 = c("res.0.6", "nUMI", "Sample"),
  datainfo.col.dog = c("res.0.6", "nUMI", "Sample"),
  cluster.col1 = "res.0.6", cluster.col2 = "res.0.6",
  cluster.col.dog = "res.0.6", withsingledog = F,
  dog.seurat.ob = NULL, neighborwithdog.seurat.ob = NULL)
}
\arguments{
\item{binary.primary}{This is the primary clustering result that can be retrieved from Tree.build.prepare object. (Because Tree.build.prepare mainly do the primary clustering.)The input here is tree.prep.ob$primary.total}

\item{datainfo.col1}{default is c("res.0.6","nUMI","Sample")}

\item{datainfo.col2}{default is c("res.0.6","nUMI","Sample")}

\item{datainfo.col.dog}{c("res.0.6","nUMI","Sample")}

\item{cluster.col1}{default is "res.0.6"}

\item{cluster.col2}{default is "res.0.6"}

\item{cluster.col.dog}{"res.0.6"}

\item{withsingledog}{Default is F}

\item{dog.seurat.ob}{=NULL}

\item{neighborwithdog.seurat.ob}{=NULL}
}
\value{
This will return an object used for Tree.build.2nd.treemaking
}
\description{
This is an important internal function to calculate distance matrix, it is used particularly in Tree.build.2nd.treemaking
}
\examples{
 	total.dist.matrxes<-dist.matrix.prep(primary.total,datainfo.col1=c(paste("res.",first.reso[1],sep=""),"nUMI","Sample"),cluster.col1=paste("res.",first.reso[1],sep=""),datainfo.col2=c(paste("res.",first.reso[2],sep=""),"nUMI","Sample"),cluster.col2=paste("res.",first.reso[2],sep=""),res1=paste("res.",first.reso[1],sep=""),res2=paste("res.",first.reso[2],sep=""))
}
