% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Do_heatmap.R
\name{Do_heatmap}
\alias{Do_heatmap}
\title{Do_heatmap}
\usage{
Do_heatmap(
  bindata,
  df1,
  df2,
  rankname,
  top_gene_num = 10,
  hardadd = NULL,
  last = T,
  insertinto = 0,
  doreturn = T
)
}
\arguments{
\item{bindata, }{"RepACT.2nd.ob$bin.data" in the repact.scRNA output list}

\item{df1, }{"RepACT.2nd.ob$BINlinear.result.summarized$UP" in the repact.scRNA output list}

\item{df2, }{"RepACT.2nd.ob$BINlinear.result.summarized$DOWN" in the repact.scRNA output list}

\item{top_gene_num, }{the number of differentially genes across peusoindex for heatmap visualization}
}
\value{
the output files will be written to disk
}
\description{
This function is to generate the heatmap for variable genes across HT->T2D 20 trajactory bins
}
\examples{
Do_heatmap(T2D.scRNA.RePACT$RepACT.2nd.ob$bin.data,df1=T2D.scRNA.RePACT$RepACT.2nd.ob$BINlinear.result.summarized$UP,df2=T2D.scRNA.RePACT$RepACT.2nd.ob$BINlinear.result.summarized$DOWN,rankname="rank",top_gene_num=20)
}
