% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreCCAtree.R
\name{DOCLEAN}
\alias{DOCLEAN}
\title{DOCLEAN}
\usage{
DOCLEAN(prep, hiorlo = "hi", res.tokeep = c(0.3, 0.6))
}
\arguments{
\item{prep}{This is the object from Tree.build.prepare}

\item{hiorlo}{if == "hi", then go delete the files in hi resolution folder. if == "lo" then go to low resolution folder to delete}

\item{res.tokeep}{This is th3e final resolution to keep. for example c(0.3,0.6)}
}
\value{
No return
}
\description{
This is to delete the past files with not-so-good resolution.
}
\examples{
S4_S5ROCK.tree.2nd.primary_0.3_0.3.list<-Second.primary.adjust(S4_S5ROCK.tree.2nd.primary_0.3_0.3.list,S4_S5ROCK.tree.prep,second.reso.adjust=c(0.3,0.6))
}
