% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{docluster.multi}
\alias{docluster.multi}
\title{docluster.multi}
\usage{
docluster.multi(Number, txcutoff = 500, sets, nms, selected = NULL,
  filterstuff = NULL, reso = 0.6, israw = F, dofindcluster = T,
  dotsne = T, acphi = 2500)
}
\arguments{
\item{Number}{the number of informative genes for clustering, default is 500}

\item{txcutoff}{the informative gene list from GetinformativeGene, default is 500}

\item{sets}{a list of dge filesw}

\item{nms}{a vector of names for each sample}

\item{selected}{NULL}

\item{filterstuff}{NULL}

\item{reso}{The resolusion for clustering,  default is 0.6}

\item{israw, }{the dge in sets is from raw raw data that has no rowname if TRUE, or if FALSE it is extracted raw matrix from Seurat object}

\item{dofindcluster}{if TRUE, do find clustering, default is true}

\item{dotsne}{if TRUE, do tsne, default is true}

\item{acphi}{accepted hiest gene number per cell, default is 2500(Should increase when processing fludigam data)}
}
\value{
this will return the clean and square plot
}
\description{
This function is to do clustering analysis for a number of sets combined
}
\examples{
S7rock_1.S7<-docluster.multi(500,txcutoff=500,sets=list(s7.RockII_1=s7.RockII_1.dge,s7.B=s7.B.dge),nms=c("s7.RockII_1","s7.B"),israw=T)
}
