% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{docluster}
\alias{docluster}
\title{docluster}
\usage{
docluster(Istopcells, mylist, name, reso = 0.6)
}
\arguments{
\item{Istopcells}{The formatted dge from dgepreprocess}

\item{mylist}{the informative gene list from GetinformativeGene}

\item{name}{a customized name for this clustering object}

\item{reso}{the clustering resolusion, default=0.6}
}
\value{
this will return the clean and square plot
}
\description{
This function is to do clustering analysis for one dataset, given a clustering resolusion
}
\examples{
S7rock_1.ob<-docluster(dgepreprocess(s7.RockII_1.dge,500,norowname=T),GetinformativeGene(dgepreprocess(s7.RockII_1.dge,500,norowname=T),500),"s7.RockII_1",reso=0.6)
}
