% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{docluster.single}
\alias{docluster.single}
\title{docluster.single}
\usage{
docluster.single(Number, Mergedset, nm1 = "", dict = NULL,
  reso = 0.6, genemin = 200, cellmin = 3)
}
\arguments{
\item{Number}{number of the most informative genes for clustering, default is 500}

\item{Mergedset}{The dge matrix}

\item{nm1}{the name for this sample}

\item{dict}{if there is any sample information, use it!}

\item{reso}{the resolution for clustering, the default is 0.6, change it accordingly}

\item{cellmin}{the minimum cell number that express a specific gene as  hard filtering for an expressed gene}

\item{geneminThe}{the minimum gene number as hard filtering for a valid cell}
}
\value{
this will return a serurat object for other analysis
}
\description{
This function a simple function to do clustering for one dge file
}
\examples{
GetinformativeGene(dgepreprocess(s7.RockII_1.dge,500,norowname=T),500)
}
