% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DonorWise.scRNA.RePACT.R
\name{DonorWise.scRNA.RePACT}
\alias{DonorWise.scRNA.RePACT}
\title{DonorWise.scRNA.RePACT}
\usage{
DonorWise.scRNA.RePACT(OBJ, pheno)
}
\arguments{
\item{OBJ, }{a scRNA-seq Seurat object}

\item{pheno, }{the column name of the "characteristics to compare" in the phenodic.use dataframe}

\item{PCrange, }{Specified PCs for regression}
}
\value{
The function return logistic model
}
\description{
This function is to run logistic regression based on the number of LSIs, and characteristics of samples.
}
\examples{
donorWise.T2D.RePACT <- DonorWise.scRNA.RePACT(OBJ=scRNA.OBJ, pheno="diseaseStat", outputname="T2D.RePACT.donorWise")
}
