% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DonorWise.snATAC.RePACT.R
\name{DonorWise.snATAC.RePACT}
\alias{DonorWise.snATAC.RePACT}
\title{DonorWise.snATAC.RePACT}
\usage{
DonorWise.snATAC.RePACT(OBJ, pheno)
}
\arguments{
\item{OBJ, }{a scRNA-seq Seurat object}

\item{pheno, }{the column name of the "characteristics to compare" in the phenodic.use dataframe}
}
\description{
This function is to run logistic regression based on the number of LSIs, and characteristics of samples.
}
