% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{format2}
\alias{format2}
\title{format2}
\usage{
format2(p, data, x, y, nolegend = T)
}
\arguments{
\item{p}{a ggplot plot}

\item{data}{the data ggplot took for plotting}

\item{x}{the column name of the x}

\item{y}{the column name of the y}

\item{nolegend}{if true then take out the legend}
}
\value{
this will return the clean and square plot
}
\description{
This function is to make a ggplot figure into a square and clean
}
\examples{
Gettopgenes(XX.ob,number)
}
