% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fullplot_v2.R
\name{Fullplot_v2}
\alias{Fullplot_v2}
\title{Fullplot_v2}
\usage{
Fullplot_v2(object, name, topgene = NULL, resolusion = "res.0.6",
  signiture = c("INS", "GCG", "SST", "PPY", "KRT19", "COL1A2", "REG1A",
  "DNAJB1", "GHRL"), p.tsnecluster = T, Pheatmap = T,
  p.tsnesample = T, p.pcacluster = T, p.pcasample = T,
  darwPCdrive = T, cell.use = 500, color = "Paired", PC34 = F,
  dotsize = 0.1, heatmapannosize = 0.5, doreturn = F, pdfwidth = 7,
  pdfheight = 7)
}
\arguments{
\item{object}{The seurat object that has been analyzed.}

\item{name}{the name for pdf file}

\item{topgene}{An optional input of cluster marker genes if "Gettopgenes" has been run outside. This can help to save some time}

\item{resolusion}{This is clustering resolution used in clustering, which can be found in object@data.info}

\item{signiture}{a vector of gene names that are intersting to be checked on tsne by color intensity}

\item{p.tsnecluster}{if true,  plot the tsne figure stained by clusters}

\item{Pheatmap}{if true,  plot the marker gene heatmap}

\item{p.tsnesample}{if true,  plot the tsne figure stained by samples}

\item{p.pcacluster}{if true,  plot the pca12 figure stained by clusters}

\item{p.pcasample}{if true,  plot the pca12 figure stained by sample}

\item{darwPCdrive}{if true,  plot the heatmap showing the pca driving genes}

\item{cell.use}{the cell number used to draw PCA driving gene plot.  Default is 500}

\item{color}{The color palette used for tsne and pca, default is "Paired"}

\item{PC34}{If ture, also plot PCA34 additionally, default is FALSE}

\item{dotsize}{The dot size for tsne and pca}

\item{doreturn}{If ture, retun plot to a variable}

\item{pdfwidth}{The width of pdf file.  Default is 7}

\item{pdfheight}{The height of pdf file.  Default is 7}
}
\value{
generate a pdf,  if  doreturn=T then return the figures
}
\description{
This function plot basic clustering result for seurat object
}
\examples{
Fullplot_v2(S7rock_1.ob,"S7rock_1.ob.pdf",topgene=NULL,resolusion="res.0.6",signiture=c("INS","GCG","SST","PPY","KRT19","COL1A2","REG1A","DNAJB1","GHRL"))
}
