% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{GetinformativeGene}
\alias{GetinformativeGene}
\title{GetinformativeGene}
\usage{
GetinformativeGene(Istopcells, Number, cutoff = 0.95)
}
\arguments{
\item{Istopcells}{The formatted dge from dgepreprocess}

\item{Number}{number of genes for out put}

\item{cutoff}{the percentile cutoff for defining a variable gene, default=0.95}
}
\value{
this will return a vector of informative genes for clustering
}
\description{
This function is to identify the most informative genes for clustering analysis
}
\examples{
GetinformativeGene(dgepreprocess(s7.RockII_1.dge,500,norowname=T),500)
}
