% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{GetPCAcelldata_v2}
\alias{GetPCAcelldata_v2}
\title{GetPCAcelldata_v2}
\usage{
GetPCAcelldata_v2(object)
}
\arguments{
\item{object}{The seurat object that has been analyzed.}
}
\value{
return a data frame with both sample/cell information and PCA information
}
\description{
This function is to quickly extract data info together with  PCA
}
\examples{
Gettopgenes(XX.ob,number)
}
