% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fullplot_v2.R
\name{GetPCheat}
\alias{GetPCheat}
\title{GetPCheat}
\usage{
GetPCheat(object, pc, cell = 500)
}
\arguments{
\item{object}{The seurat object that has been analyzed.}

\item{pc}{the PC#  to plot.  input could be 1 or 2 or 3 ...}

\item{cell}{A number of cells to be used for plotting}
}
\value{
Return the figure
}
\description{
This function plot heatmap for PCA driving genes
}
\examples{
pc1.driver<-GetPCheat(object,1,cell=cell.use)+ggtitle("PC1")
}
