% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRePACTmodel.cca.R
\name{GetRePACTmodel.cca}
\alias{GetRePACTmodel.cca}
\title{GetRePACTmodel.cca}
\usage{
GetRePACTmodel.cca(
  ccaWithinfo = cca.L2.info,
  prefix = "CC",
  pheno = "Disease",
  CCrange = 1:10
)
}
\arguments{
\item{ccaWithinfo, }{phenotable}

\item{pheno, }{the column name of the "characteristics to compare" in the phenodic.use dataframe}

\item{PCrange, }{Specified PCs for regression}
}
\value{
The function return logistic model
}
\description{
This function is to run logistic regression based on the number of LSIs, and characteristics of samples.
}
\examples{
RepACT.obj<-Prepareforpseudoregress.g(OBJ,PCrange=PCrange,phenodic.use=sub_phenotable,pheno=pheno,linear=T)
}
