% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gettopgenes.R
\name{Gettopgenes}
\alias{Gettopgenes}
\title{Gettopgenes}
\usage{
Gettopgenes(object, number)
}
\arguments{
\item{object}{The seurat object that has been analyzed.}

\item{number}{number of genes fro each cluster}
}
\value{
generate a pdf,  if  doreturn=T then return the figures
}
\description{
This function is based on seurat function to generate marker genes for each cluster
}
\examples{
Gettopgenes(XX.ob,number)
}
