% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GettsnesignatureSuper.R
\name{GettsnesignatureSuper}
\alias{GettsnesignatureSuper}
\title{GettsnesignatureSuper}
\usage{
GettsnesignatureSuper(object, object.all, signiture = c("INS", "RBP4",
  "FFAR4", "ID1", "ID2", "ID3", "DNAJB1"), doPCA = F, dotSNE = T,
  usePC34 = F, extratitle = "", toreturn = F, dotsize = 0.3,
  buttomgrey = T, nolegend = T, highcolor = "red")
}
\arguments{
\item{object}{The seurat object that has been analyzed.}

\item{object.all}{The total object for normalization.  Usually it is the same as   object}

\item{signiture}{a vector of gene names that are intersting to be checked on tsne by color intensity}

\item{doPCA}{if ture then stain on PCA12 plot.  default is F}

\item{dotSNE}{if ture then stain on tsne plot.  default is T}

\item{usePC34}{if true,  plot the PC34 istead of PC12. Default is F}

\item{extratitle}{A string that can be added to the title of figures. default is ""}

\item{dotsize}{The dot size for the figures,  default is 0.3}

\item{buttomgrey}{if true,  The cells with no expression set color as grey}

\item{nolegend}{if true,  no color intensity legend is put}

\item{highcolor}{The color for highest expression, A color spectrum is formed between "grey" to the color set. Default is red}
}
\value{
draw figures. if toreturn=T, then return a list of figures
}
\description{
This function plot tsne or PCA stained by gene relative expression
}
\examples{
Fullplot_v2(S7rock_1.ob,"S7rock_1.ob.pdf",topgene=NULL,resolusion="res.0.6",signiture=c("INS","GCG","SST","PPY","KRT19","COL1A2","REG1A","DNAJB1","GHRL"))
}
