% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InSample.compare.ATAC.nb.R
\name{InSample.compare.ATAC.nb}
\alias{InSample.compare.ATAC.nb}
\title{InSample.compare.ATAC.nb}
\usage{
InSample.compare.ATAC.nb(
  matrix = ALL.Beta.ATAC.Mtx,
  meta = Beta.all.metadata,
  pseudo = beta.atac.LSI.withinfo[, "pseudo.index.balanced", drop = F],
  sample.col = "pubdonor",
  peak = "chr11_92949418_92950719",
  GetDatatoplot = T
)
}
\arguments{
\item{matrix, }{cellxpeak matrix}
}
\value{
The function return logistic model
}
\description{
This function is to identify intra-donor genes by negative binomial
}
\examples{
donorWise.T2D.RePACT <- DonorWise.scRNA.RePACT(OBJ=scRNA.OBJ, pheno="diseaseStat", outputname="T2D.RePACT.donorWise")
}
