% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeEvenBinBydepth.R
\name{MakeEvenBinBydepth}
\alias{MakeEvenBinBydepth}
\title{MakeEvenBinBydepth}
\usage{
MakeEvenBinBydepth(
  cellvsPeak.m = BetacellvsPeak.m.filtered,
  data.info = BetaPeak.data.info,
  binnumber = 20
)
}
\arguments{
\item{cellvsPeak.m, }{}
}
\value{
The function return a list: "PCvariance", "PCanfpheno", "object.raw.withinfo", "model", "reg.plot.2d", "model.para"
}
\description{
This function is to run regression (linear/logistic) based on the number of PCs, and characteristics of samples.
}
\examples{
RepACT.obj<-Prepareforpseudoregress.g(OBJ,PCrange=PCrange,phenodic.use=sub_phenotable,pheno=pheno,linear=T)
}
