% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_RePACT.r
\name{MakeEvenBinBydepth_SpeedUP}
\alias{MakeEvenBinBydepth_SpeedUP}
\title{MakeEvenBinBydepth}
\usage{
MakeEvenBinBydepth_SpeedUP(OBJ, data.info = BetaPeak.data.info, binnumber = 20)
}
\arguments{
\item{OBJ, }{Seurat OBJ}
}
\value{
The function return a list: # data.info.withbin:(data.info + evenfragbin), cellvsPeak.m.aggr: gene ~ traj1-traj20 , index: mean pseudoindex of every traj_bin, depth: total frags of cells within each traj_bin
}
\description{
This function is to make even bins for cells, to keep the cell number and total cell fragments even.
}
\examples{
beta.RNA.PCA.20bin.ob <- MakeEvenBinBydepth_SpeedUP(OBJ=OBJ, data.info=BetaPCA[,51:ncol(BetaPCA)], binnumber=20)
}
