% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreCCAtree.R
\name{myddply.center}
\alias{myddply.center}
\title{myddply.center}
\usage{
myddply.center(df, Col, is.originalcluster = F, combinecol = "Sample")
}
\value{
This will return an object for dist.matrix.prep and dist.matrix.prep.v3
}
\description{
This is an internal function used by dist.matrix.prep and dist.matrix.prep.v3
}
\examples{
Sample1.centers<-myddply.center(ob1.PCAS.clst,cluster.col1)
}
