% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreCCAtree.R
\name{myddply.center.v2}
\alias{myddply.center.v2}
\title{myddply.center.v2}
\usage{
myddply.center.v2(df.1, df.2, Col1, Col2)
}
\value{
This will return an object for dist.matrix.prep and dist.matrix.prep.v3
}
\description{
This is another internal function used by dist.matrix.prep and dist.matrix.prep.v3
}
\examples{
Sample1.centers<-myddply.center(ob1.PCAS.clst,cluster.col1)
}
