% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreCCAtree.R
\name{mydplyr.percentage}
\alias{mydplyr.percentage}
\title{mydplyr.percentage}
\usage{
mydplyr.percentage(df, by = "stage")
}
\description{
This function is an internal function
}
\examples{
Primaryclutering(dgelist,txcut=500,cluster.resos=first.reso,RAWinput=israw)
}
