% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snATAC.RePACT.r
\name{plot_scATAC_RePACT_heatmap}
\alias{plot_scATAC_RePACT_heatmap}
\title{plot_scATAC_RePACT_heatmap}
\usage{
plot_scATAC_RePACT_heatmap(peakSet, RePACT_OBJ)
}
\arguments{
\item{peakSet, }{a vector of peaks (":" "-")}

\item{RePACT_OBJ, }{RePACT reesult OBJ}
}
\value{
The function return a plot
}
\description{
This function is to take a vetor of peaks and plot heatmaps from RePACT result
}
