% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_RePACT.r
\name{plot_scRNA_RePACT_heatmap}
\alias{plot_scRNA_RePACT_heatmap}
\title{plot_scRNA_RePACT_heatmap}
\usage{
plot_scRNA_RePACT_heatmap(geneset, RePACT_OBJ, ifShowGene = T)
}
\arguments{
\item{geneset, }{a vector of genes}

\item{RePACT_OBJ, }{RePACT reesult OBJ}
}
\value{
The function return a plot
}
\description{
This function is to take a vetor of genes and plot heatmaps from RePACT result
}
