% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreCCAtree.R
\name{Prep.res.adjust}
\alias{Prep.res.adjust}
\title{Prep.res.adjust}
\usage{
Prep.res.adjust(prep, first.reso.adjust = c(0.03, 0.03))
}
\arguments{
\item{prep}{the output from Tree.build.prepare}

\item{first.reso.adjust}{This is the ADJUSTED CLUSTER RESOLUTION.The format is the same as First.reso in Tree.build.prepare}
}
\value{
This will return the same object as that generated from Tree.build.prepare
}
\description{
This function is to adjust the clustering resolution generated from Tree.build.prepare
}
\examples{
 S4_S5ROCK.tree.prep<-Prep.res.adjust(S4_S5ROCK.tree.prep,first.reso.adjust=c(0.015,0.015))
}
