% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepareforpseudoregress.g.R
\name{Prepareforpseudoregress.g}
\alias{Prepareforpseudoregress.g}
\title{Prepareforpseudoregress.g}
\usage{
Prepareforpseudoregress.g(
  object = NULL,
  PCrange = 1:10,
  phenodic.use = NULL,
  pheno,
  dotsize = 0.3,
  linear = T
)
}
\arguments{
\item{object, }{A scRNA Seurat (V3/V4), with PCs and clustering info}

\item{PCrange, }{Specified PCs for regression}

\item{phenodic.use, }{a dataframe contains Sample column(it refers to donor in this study), characteristics to compare(it referes to if the donor/cells are from healthy or T2D, or it can be continuous, e.g. BMI)}

\item{pheno, }{the column name of the "characteristics to compare" in the phenodic.use dataframe}

\item{dotsize, }{the size of dots in the output plots}

\item{linear, }{if the "characteristics to compare" is binary(e.g. if T2D) or continuous(e.g. BMI)}
}
\value{
The function return a list: "PCvariance", "PCanfpheno", "object.raw.withinfo", "model", "reg.plot.2d", "model.para"
}
\description{
This function is to run regression (linear/logistic) based on the number of PCs, and characteristics of samples.
}
\examples{
RepACT.obj<-Prepareforpseudoregress.g(OBJ,PCrange=PCrange,phenodic.use=sub_phenotable,pheno=pheno,linear=T)
}
