% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreCCAtree.R
\name{Primaryclutering}
\alias{Primaryclutering}
\title{Primaryclutering}
\usage{
Primaryclutering(rawdatapackage, txcut = 500, RAWinput = T,
  cluster.resos = c(0.6, 0.6), allsamereso = F, together.reso = 0.6)
}
\arguments{
\item{rawdatapackage}{a list of several dge files. At least 2 dge. in the case of TreCCA tree. It is alwasy two dge, one from early time, one form latert}

\item{txcut}{The transcripts cutoff for clustering analysis}

\item{RAWinput}{If true, it is a raw dge}

\item{cluster.resos}{A vector of number, each is the cluster resolution for each dge.}

\item{allsamereso}{If true, then take the first number of cluster.resos for clustering of all dges}

\item{Together.reso}{The resolution of merged dge(all samples together)}
}
\value{
This will return a list of seurat result. It is used in function such as Tree.build.prepare
}
\description{
This function is an advanced clustering function that when for example input dgeA, dgeB. This will do cluster individually for A, B and also for merged AB
}
\examples{
Primaryclutering(dgelist,txcut=500,cluster.resos=first.reso,RAWinput=israw)
}
