% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreCCAtree.R
\name{Refine.relationship}
\alias{Refine.relationship}
\title{Refine.relationship}
\usage{
Refine.relationship(dist.matrix = deeper.dist.matrxes.current,
  all.matrix = p.dendro[[length(p.dendro)]]$matrix)
}
\arguments{
\item{dist.matrix}{deeper.dist.matrxes.current}

\item{all.matrix}{p.dendro[[length(p.dendro)]]$matrix}
}
\value{
This will return an object used in Tree.build.2nd.treemaking.  return the new "XX.dist.matrxes.current",  assign it back
}
\description{
This is and internal function used in Tree.build.2nd.treemaking to refine the relationship if one upstreme nod is connected to more than one nods in the downstreme stage.  Input is Input\    deeper.dist.matrxes.current  and p.dendro
}
\examples{
Refine.relationship(deeper.dist.matrxes.current,p.dendro[[length(p.dendro)]]$matrix)
}
