% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repact.scRNA.R
\name{repact.scRNA}
\alias{repact.scRNA}
\title{repact.scRNA.R}
\usage{
repact.scRNA(
  OBJ,
  PCrange = 1:20,
  pheno,
  is_continuous,
  norm_index = T,
  SlopeCut = 0,
  output_name2
)
}
\arguments{
\item{OBJ, }{A scRNA Seurat (V3/V4), with PCs and clustering info, OBJ@meta.data should contain "Sample"(donor) column and the column of the chracteristics to compare, e.g. if_T2D, BMI}

\item{PCrange, }{Specified PCs for regression}

\item{pheno, }{the column name of the "characteristics to compare" in the OBJ@meta.data}

\item{is_continuous, }{if the "characteristics to compare" is continuous}

\item{norm_index, }{if the pseudo index should be normalized}

\item{SlopeCut, }{the cutoff of slope to define differentially genes across pseudo index (n=binnumber)}

\item{output_name2, }{a path/outputname to write out the results to disk, including .rds, .csv, .pdf}
}
\value{
.rds, .csv, .pdf to disk
}
\description{
This function is to run repact on a Seurat object.
}
\examples{
repact.scRNA.R(OBJ, PCrange=1:20,pheno="diseaseStat",is_continuous="F",norm_index=T,SlopeCut=0.05,output_name2="T2D_Beta.scRNA.RePACT")
}
