% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repact_logistic.snATAC.R
\name{repact_logistic.snATAC}
\alias{repact_logistic.snATAC}
\title{repact_logistic.snATAC}
\usage{
repact_logistic.snATAC(OBJ, LSIrange, pheno, outputname)
}
\arguments{
\item{OBJ, }{A snATAC Seurat (V3/V4), with LSIs and clustering info, the metadata should contain Sample column(it refers to donor in this study), characteristics to compare(it referes to if the donor/cells are from healthy or T2D, or it can be continuous, e.g. BMI)}

\item{LSIrange, }{Specified PCs for regression}

\item{pheno, }{the column name of the "characteristics to compare" in the phenodic.use dataframe}

\item{outputname, }{the prefix of output files}
}
\value{
The function will write output files .pdf, .rds to disk
}
\description{
This function is to run regression (linear/logistic) based on the number of PCs, and characteristics of samples.
}
\examples{
repact_logistic.snATAC(snATAC.OBJ, LSIrange=1:20, pheno="diseaseStat", outputname="T2D_Beta.snATAC.RePACT")
}
