% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_RePACT.r
\name{scRNA.RePACT}
\alias{scRNA.RePACT}
\title{scRNA.RePACT optimize speed}
\usage{
scRNA.RePACT(
  OBJ,
  Sample,
  pheno,
  pheno_levels,
  is_continuous = F,
  if_donorWise = F,
  binnumber = 20,
  PCrange = "",
  RePACT_qvalCut = 0.005,
  donorWise_qvalCut = 0.01
)
}
\arguments{
\item{OBJ, }{a scRNA-seq Seurat object}

\item{Sample, }{colnames of donor or sample infomation in OBJ@meta.data}

\item{pheno, }{the column name of the "characteristics to compare" in OBJ@meta.data, e.g. diseaseStatus or BMI or HBA1C}

\item{pheno_levels, }{specify the levels of pheno column. e.g. c("HT", "T2D"), specify for binary pheno}

\item{is_continuous, }{if pheno is continous variable. default is F. e.g. diseaseStatus is F, BMI is T}

\item{if_donorWise, }{if perform donor wise RePACT, default is F}

\item{binnumber, }{number of bins for cell grouping}

\item{RePACT_qvalCut, }{qvalue cutoff to determine the significant genes along the disease or other characteristics trajactory, certain cutoff for slope can be further applied}

\item{donorWise_qvalCut, }{qvalue cutoff to determine the significant genes varying within donors or across donors.}

\item{PCrange.}{default is "", automatically take top 10 PCs that are significant with phenotype, otherwise specify 10 PCs, e.g. 1:10 or 2:11}
}
\value{
The function return a list of objects
}
\description{
This function is to run regression based on the number of PCs, and characteristics of samples.
}
\examples{
T2D.scRNA.RePACT <- scRNA.RePACT(OBJ=scRNA.OBJ,Sample="Donor", pheno="diseaseStat", is_continuous=F, if_donorWise=F)
}
