% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreCCAtree.R
\name{Second.primary.adjust}
\alias{Second.primary.adjust}
\title{Second.primary.adjust}
\usage{
Second.primary.adjust(second.pr, prep, second.reso.adjust = c(0.3, 0.6))
}
\arguments{
\item{second.pr}{This is the secondary clustering object from Tree.build.2nd.clustering}

\item{prep}{this is the low resolution clutsering object. from Tree.build.prepare}

\item{second.reso.adjust}{This is the adjusted resolution    for example: c(0.3,0.6)}
}
\value{
This will return.   The adjusted secondary clustering, assign back to the same object generated from Tree.build.2nd.clustering
}
\description{
This is to adjust the clustering resolution for secondary clusters.
}
\examples{
S4_S5ROCK.tree.2nd.primary_0.3_0.3.list<-Second.primary.adjust(S4_S5ROCK.tree.2nd.primary_0.3_0.3.list,S4_S5ROCK.tree.prep,second.reso.adjust=c(0.3,0.6))
}
