% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_RePACT.r
\name{SelectSigPCs}
\alias{SelectSigPCs}
\title{SelectSigPCs}
\usage{
SelectSigPCs(BetaPCA, Sample, pheno, is_continuous = F)
}
\arguments{
\item{BetaPCA, }{a dataframe, combined PCs and meta data}

\item{Sample}{sample/donor name}

\item{pheno}{phenotype to compare, binary e.g. diseaseStat; continuous e.g. BMI}

\item{is_continuous}{T or F, if pheno is cpontinuous}
}
\value{
a dataframe: PC_number, PC, PC_pval, ordered by PC_pval
}
\description{
This function is to run regression (linear/logistic) based on the number of PCs, and characteristics of samples.
}
\examples{
sigPCs <- SelectSigPCs(BetaPCA, Sample, pheno, is_continuous=F)
}
