% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snATAC.RePACT.r
\name{snATAC.RePACT}
\alias{snATAC.RePACT}
\title{snATAC.RePACT}
\usage{
snATAC.RePACT(
  OBJ,
  Sample,
  pheno,
  pheno_levels,
  is_continuous = F,
  if_donorWise = F,
  binnumber = 20,
  LSIrange = "",
  RePACT_qvalCut = 0.01,
  donorWise_qvalCut = 0.01
)
}
\arguments{
\item{OBJ, }{a snATAC-seq Seurat object, assay name is "peak"}

\item{Sample, }{colnames of donor or sample infomation in OBJ@meta.data}

\item{pheno, }{the column name of the "characteristics to compare" in OBJ@meta.data, e.g. diseaseStatus or BMI or HBA1C}

\item{pheno_levels, }{specify the levels of pheno column. e.g. c("HT", "T2D"), specify for binary pheno}

\item{is_continuous, }{if pheno is continous variable. default is F. e.g. diseaseStatus is F, BMI is T}

\item{if_donorWise, }{if perform donor wise RePACT, default is F}

\item{binnumber, }{number of bins for cell grouping}

\item{RePACT_qvalCut, }{qvalue cutoff to determine the significant genes along the disease or other characteristics trajactory, certain cutoff for slope can be further applied}

\item{donorWise_qvalCut, }{qvalue cutoff to determine the significant genes varying within donors or across donors.}

\item{LSIrange.}{default is "", automatically take top 10 LSIs that are significant with phenotype, otherwise specify 10 LSIs, e.g. 1:10 or 2:11}
}
\value{
, a list of objects
}
\description{
This function is to run logistic regression based on the number of LSIs, and characteristics of samples.
}
