% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA_RePACT.r
\name{standard_fun}
\alias{standard_fun}
\title{standard_fun}
\usage{
standard_fun(OBJ)
}
\arguments{
\item{OBJ, }{Seurat OBJ}
}
\value{
The function return a Seurat OBJ
}
\description{
This function is to run basic processing for Seurat OBJ
}
