% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RePACT.R
\name{Tjct.core.plot}
\alias{Tjct.core.plot}
\title{Tjct.core.plot}
\usage{
Tjct.core.plot(object = NULL, secondobj = NULL, pheno = NULL,
  f1.name = "XX.10d.violin.pdf", f2.name = "XX.his.pdf",
  f3.name = "XX.trj.heatmap.pdf", f3.height = 12,
  f3.tittle = "cell type:Changing genes on phenotype trajectory\\ntop6\%",
  table1.name = "XX.traj.up.genes-q0.05top0.06.csv",
  table2.name = "XX.traj.dowb.genes-q0.05top0.06.csv", rankcut = 0.06,
  colorset = "Set1", do.return = F)
}
\arguments{
\item{object}{The first trajectory generated by Prepareforpseudoregress.g}

\item{secondobj}{The regressed and bined object by Tjct.core.gen}

\item{pheno}{This is the column we are interested for RePACT gotta be the same with previous Prepareforpseudoregress.g}

\item{f1.name}{The pdf name of violin plot "XX.10d.violin.pdf".  10 d means 10 PCs used.}

\item{f2.name}{The pdf name of histgram "XX.his.pdf".}

\item{f3.name}{The pdf name of the heatmap "XX.trj.heatmap.pdf".}

\item{f3.height}{The pdf heigh for heatmap, change this when the number of genes shown is changed, default is 12}

\item{f3.tittle}{The figure title for heatmap."cell type:Changing genes on phenotype trajectory\ntop6%"}

\item{table1.name}{The name of a csv file  for upregulated genes"XX.traj.up.genes-q0.05top0.06.csv"}

\item{table2.name}{The name of a csv file  for upregulated genes"XX.traj.up.genes-q0.05top0.06.csv"}

\item{rankcut}{The percentile to be shown on the heatmap. default is 0.06}

\item{colorset}{The color pallete"Set1" "Set2" "Set3"}
}
\value{
A list of upregulated and downregulated trajectory genes.
}
\description{
This function is to generate major plots for Repact analysis
}
\examples{
Tjct.core.plot(BMI.tjct.ob,BMI.tjct.2nd.ob,pheno="BMI",f1.name="BMI.tjct.10d.violin.pdf",f2.name="BMI.tjct.his.pdf",f3.name="BMI.tjct.trj.heatmap.pdf",f3.height=14,f3.tittle="cell type:Changing genes on phenotype trajectory\\ntop6\%",table1.name="BMI.tjct.traj.up.genes-q0.05Full.csv",table2.name="BMI.tjct.traj.dowb.genes-q0.05Full.csv",rankcut=0.05)
}
