% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tjct.core.plot.ss.R
\name{Tjct.core.plot.ss}
\alias{Tjct.core.plot.ss}
\title{Tjct.core.plot.ss}
\usage{
Tjct.core.plot.ss(
  object = NULL,
  secondobj = NULL,
  phenodic.use = NULL,
  pheno = NULL,
  top_gene_num = 10,
  output_name,
  norm_index = F
)
}
\arguments{
\item{object, }{output of Prepareforpseudoregress.g function}

\item{secondobj, }{output of Tjct.core.gen function}

\item{phenodic.use, }{a dataframe contains Sample column(it refers to donor in this study), characteristics to compare(it referes to if the donor/cells are from healthy or T2D, or it can be continuous, e.g. BMI)}

\item{pheno, }{the column name of the "characteristics to compare" in the phenodic.use dataframe}

\item{top_gene_num, }{the number of differentially genes across peusoindex for heatmap visualization}

\item{output_name, }{a path/outputname to write out the results to disk, including .rds, .csv, .pdf}

\item{norm_index, }{if the pseudo index should be normalized}
}
\value{
the output files will be written to disk
}
\description{
This function is to generate the PC scatterplots, Sample-wise pseudo Index violin plot, heatmaps for specified number of differentially genes across peusoindex.
}
\examples{
Tjct.core.plot.ss(RepACT.obj,RepACT.2nd.ob,pheno=pheno,phenodic.use=sub_phenotable,top_gene_num=15,output_name=output_name2,norm_index=T)
}
