% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{Tomerge}
\alias{Tomerge}
\title{Tomerge}
\usage{
Tomerge(A, B)
}
\arguments{
\item{A}{dataframe A}

\item{B}{dataframe B}
}
\value{
return a data frame with merged information
}
\description{
This function is to quickly merge two dataframe by rownames, but only leave the rows both dataframe have
}
\examples{
Gettopgenes(XX.ob,number)
}
