% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreCCAtree.R
\name{Tree.build.1}
\alias{Tree.build.1}
\title{Tree.build.1}
\usage{
Tree.build.1(tree.prep.ob)
}
\arguments{
\item{tree.prep.ob}{This is an object generated from Tree.build.prepare or Prep.res.adjust}
}
\value{
This will return an object used for Tree.build.1
}
\description{
This is the subtree relationship builder for low resolution
}
\examples{
 	S4_S5ROCK.tree.1.ob<-Tree.build.1(S4_S5ROCK.tree.prep)
}
