% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreCCAtree.R
\name{Tree.build.2nd.clustering.patch}
\alias{Tree.build.2nd.clustering.patch}
\title{Tree.build.2nd.clustering.patch}
\usage{
Tree.build.2nd.clustering.patch(primaries.deeper.lst, tree.prep, tree.1.ob,
  singledog.reso = 0.3)
}
\arguments{
\item{primaries.deeper.lst}{This is generated by Tree.build.2nd.clustering and Second.primary.adjust}

\item{tree.prep}{This is the low resolution clustering result from Tree.build.prepare}

\item{tree.1.ob}{This is the low reslution subtree result from Tree.build.1}

\item{singledog.reso}{This is the resolution singledog cluster for 0.3}
}
\value{
This will return the high resolution result with a pach that consider the single dog.  This is going to be used for Tree.build.2nd.treemaking
}
\description{
Introduction\   This function is to address single dog problem. It will Identify and do seurat clustering for the single dog & dog+neighbours. ### Input\    the clustered result from Tree.build.2nd.clustering as well as tree.prep  and tree.1.ob. singledog.reso should be specified. ### Output\     If there was no single dog, the out put is exactly the same as imput,  otherwise, two seurat object would be added at the end, forming a list as whole
}
\examples{
S4_S5ROCK.tree.2nd.primary_0.3_0.3.list<-Tree.build.2nd.clustering(S4_S5ROCK.tree.prep,S4_S5ROCK.tree.1.ob,second.reso=c(0.3,0.3))
}
