% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreCCAtree.R
\name{Tree.build.2nd.clustering}
\alias{Tree.build.2nd.clustering}
\title{Tree.build.2nd.clustering}
\usage{
Tree.build.2nd.clustering(tree.prep, tree.1.ob, second.reso = c(0.3,
  0.3))
}
\arguments{
\item{tree.prep}{This is an object generated from Tree.build.prepare or Prep.res.adjust.  This is basically the clustering result.}

\item{tree.1.ob}{This is an object generated from Tree.build.1.  This is the low resulotion subtree}

\item{second.reso}{This is the resolution for the high resolution.  For example, c(0.3,0.3)}
}
\value{
This will return the secondary clustering result. It will be used for Tree.build.2nd.clustering.patch and Tree.build.2nd.treemaking
}
\description{
This function did seurat primary clustering for the second layer,  which is higher resolution . Input\   tree.prep,tree.1.ob, which is the output from Tree.build.1;  thesecondary resolution for clustering. ### Output\   primaries.deeper.lst which is a list primary objects, including multiple pairs of connections.
}
\examples{
S4_S5ROCK.tree.2nd.primary_0.3_0.3.list<-Tree.build.2nd.clustering(S4_S5ROCK.tree.prep,S4_S5ROCK.tree.1.ob,second.reso=c(0.3,0.3))
}
