% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreCCAtree.R
\name{Tree.build.2nd.treemaking}
\alias{Tree.build.2nd.treemaking}
\title{Tree.build.2nd.treemaking}
\usage{
Tree.build.2nd.treemaking(primaries.deeper.lst, second.reso = c(0.3,
  0.3), singledog.reso = 0.06, downstremename = "s1.24",
  upstremename = "s1", relationtext = 2.5, dorefine = T, dir)
}
\arguments{
\item{primaries.deeper.lst}{This is the adjusted secondary  clustering results from Tree.build.2nd.clustering.patch}

\item{second.reso}{This should be  consistent with what eventually used in secondary clustering,  it should follow Second.primary.adjust}

\item{singledog.reso}{This should follow that in Tree.build.2nd.clustering.patch}

\item{downstremename}{="s1.24"}

\item{upstremename}{="s1"}

\item{relationtext}{=2.5}

\item{dorefine}{=T}

\item{dir}{The path of the whole tree generating.  keep consistent with Tree.build.prepare}
}
\value{
No return
}
\description{
Introduction\   Similar with Tree.build.1,  this function after getting the clustering primary object for the second layer. this one is literally building the tree and generate plots
}
\examples{
S4_S5ROCK.tree.2nd.primary_0.3_0.3.list<-Second.primary.adjust(S4_S5ROCK.tree.2nd.primary_0.3_0.3.list,S4_S5ROCK.tree.prep,second.reso.adjust=c(0.3,0.6))
}
