% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreCCAtree.R
\name{Tree.build.prepare}
\alias{Tree.build.prepare}
\title{Tree.build.prepare}
\usage{
Tree.build.prepare(dge1, dge2, name1, name2, first.reso = c(0.03, 0.03),
  treemake.dir = "./")
}
\arguments{
\item{dge1}{the dge file (usually from earlier timepoint), either raw or processed is acceptable}

\item{dge2}{the dge file (usually from later timepoint), either raw or processed is acceptable}

\item{name1}{The name for the dge1 (usually from earlier timepoint)}

\item{name2}{The name for the dge2 (usually from later timepoint)}

\item{first.reso}{The cluster resolution for dge1 and dge2 respectively. A vector of two number , for example c(0.03,0.03)}

\item{treemake.dir}{A path. The workspace where the whole tree making project is on}
}
\value{
A list of stuff including seurat clutsering result; fullplots; the resolution used and  the path
}
\description{
This function is the first function to start calculate the subtree from dgeA, dgeB. This  basically do the clustering and fullplot at low resolution
}
\examples{
 S4_S5ROCK.tree.prep<-Tree.build.prepare(dge1=s4.B.dge,dge2=S5.ROCKII.dge,name1="S4.B",name2="S5.rock",first.reso=c(0.03,0.03),treemake.dir="/mnt/NFS/homeGene/JinLab/cxw486/Dropseq/Entrance/Esderived/Esdrived_DGE/AttemptFrom17.8.28/ROCKII/Treemake/")
}
