% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreCCAtree.R
\name{tree.plot.v2}
\alias{tree.plot.v2}
\title{tree.plot.v2}
\usage{
tree.plot.v2(maturername = downstremename, youngername = upstremename,
  withsingledog = F, dendrodata,
  matrixes = deeper.dist.matrxes.current, print.later.names = T,
  print.earlier.names = T, print.stage.level = T,
  inputorder.maturer = F, inputorder.yonger = F,
  order.maturer = NULL, order.yonger = NULL, textsize = 2,
  ballsize = 5, percent = 2, nocellnumber = F,
  thecellnumber = cellnumber)
}
\arguments{
\item{maturername}{=downstremename,}

\item{youngername}{=upstremename,}

\item{withsingledog}{=F,dendrodata,}

\item{matrixes}{=deeper.dist.matrxes.current,}

\item{print.later.names}{=T,}

\item{print.earlier.names}{=T,}

\item{print.stage.level}{=T,}

\item{inputorder.yonger}{=F,}

\item{order.maturer}{=NULL,}

\item{order.yonger}{=NULL,}

\item{textsize}{=2,}

\item{ballsize}{=5,}

\item{percent}{=2,}

\item{nocellnumber}{=F,}

\item{thecellnumber}{=cellnumber}

\item{inputorder.mature}{r=F,}
}
\value{
This will return an object used in Tree.build.2nd.treemaking.  return the new "XX.dist.matrxes.current",  assign it back
}
\description{
This is and internal function used in Tree.build.2nd.treemaking To plot the tree
}
\examples{
Refine.relationship(deeper.dist.matrxes.current,p.dendro[[length(p.dendro)]]$matrix)
}
