% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallT2Dpeak_pvalueOneTail.R
\name{CallT2Dpeak_pvalueOneTail}
\alias{CallT2Dpeak_pvalueOneTail}
\title{CallT2Dpeak_pvalueOneTail}
\usage{
CallT2Dpeak_pvalueOneTail(
  cellvsPeak.m.aggr = ATAConCCA.betaT2D.tjct.3nd.10bin.ob$cellvsPeak.m.aggr,
  depths = ATAConCCA.betaT2D.tjct.3nd.10bin.ob$depths,
  index = ATAConCCA.betaT2D.tjct.3nd.10bin.ob$index,
  doscale = T,
  GlobalSlopes
)
}
\description{
CallT2Dpeak_pvalueOneTail
}
