% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallT2Dpeak_qvalue.R
\name{CallT2Dpeak_qvalue}
\alias{CallT2Dpeak_qvalue}
\title{CallT2Dpeak_qvalue}
\usage{
CallT2Dpeak_qvalue(
  cellvsPeak.m.aggr = ATAConCCA.betaT2D.tjct.3nd.10bin.ob$cellvsPeak.m.aggr,
  depths = ATAConCCA.betaT2D.tjct.3nd.10bin.ob$depths,
  index = ATAConCCA.betaT2D.tjct.3nd.10bin.ob$index,
  qcut = 0.1,
  slopecut1 = 0.5,
  slopecut2 = -0.5,
  doscale = T
)
}
\arguments{
\item{cellvsPeak.m.aggr, }{A scRNA Seurat (V3/V4), with PCs and clustering info}
}
\value{
The function return a list: "PCvariance", "PCanfpheno", "object.raw.withinfo", "model", "reg.plot.2d", "model.para"
}
\description{
This function is to run regression (linear/logistic) based on the number of PCs, and characteristics of samples.
}
\examples{
RepACT.obj<-Prepareforpseudoregress.g(OBJ,PCrange=PCrange,phenodic.use=sub_phenotable,pheno=pheno,linear=T)
}
