% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tjct.core.gen.R
\name{Tjct.core.gen}
\alias{Tjct.core.gen}
\title{Tjct.core.gen}
\usage{
Tjct.core.gen(
  object = NULL,
  binnumber = 20,
  qcut = 0.05,
  norm_index = F,
  SlopeCut = 0
)
}
\arguments{
\item{object, }{output of Prepareforpseudoregress.g function}

\item{binnumber, }{the number of bins separated based on even distance on PC space}

\item{qcut, }{the cutoff of qvalue to define differentially genes across pseudo index (n=binnumber)}

\item{norm_index, }{if the pseudo index should be normalized}

\item{SlopeCut, }{the cutoff of slope to define differentially genes across pseudo index (n=binnumber)}
}
\value{
a list contains geneUMI.dic, raw.bin, bin.data, binregress, BINlinear.result, BINlinear.result.summarized, parameter, FC.result
}
\description{
This function is to spearate cells into the specified number of bins based on even distance on PC space.
}
\examples{
Tjct.core.gen(RepACT.obj,binnumber=20,norm_index=T,SlopeCut=SlopeCut)
}
